+++ ut-robomaster/src/drivers/as5600.hpp
@@ -3,15 +3,13 @@
 #include "modm/architecture/interface/i2c_device.hpp"
 #include "modm/processing/protothread/protothread.hpp"
 
-namespace encoder
+namespace driver
 {
 template <class I2cMaster>
-class Encoder : public modm::I2cDevice<I2cMaster, 10, modm::I2cWriteReadTransaction>,
-                public modm::pt::Protothread
+class As5600 : public modm::I2cDevice<I2cMaster, 1>, public modm::pt::Protothread
 {
 public:
-    Encoder(uint8_t address = slave_address)
-        : modm::I2cDevice<I2cMaster, 10, modm::I2cWriteReadTransaction>(address){};
+    As5600() : modm::I2cDevice<I2cMaster, 1>(ADDRESS){};
 
     void update() { run(); }
 
@@ -21,25 +19,24 @@ public:
 
         while (true)
         {
-            PT_WAIT_UNTIL(this->startWriteRead(writeBuf, 1, buffer, 2));
+            buffer[0] = uint8_t(Register::RAWANGLE);
+            PT_WAIT_UNTIL(this->startWriteRead(buffer, 1, buffer, 2));
             PT_WAIT_WHILE(this->isTransactionRunning());
 
             if (this->wasTransactionSuccessful())
             {
                 angle = (buffer[0] << 8) | buffer[1];
             }
-
-            PT_YIELD();
         }
 
         PT_END();
     }
 
+    /// @brief Get current measured angle of the encoder
+    /// @return Angle (revs)
     float getAngle() { return angle / 4096.0f; }
 
 protected:
-    static const uint8_t slave_address = 0x36;  // address of the AS5600
-
     enum class Register : uint8_t
     {
         ZPOS = 0x01,
@@ -55,9 +52,9 @@ protected:
     };
 
 private:
-    uint16_t angle;
+    static const uint8_t ADDRESS = 0x36;
+    uint16_t angle = 0;
     uint8_t buffer[2];
-    const uint8_t writeBuf[1] = {0x0E};
 };
 
-};  // namespace encoder
+};  // namespace driver
